+++ Engine/Source/Physics/Body/PhysicsBody.cpp
@@ -23,4 +23,13 @@ namespace Core
 
         return owner->GetTransform()->Position;
     }
+
+    Vector3 PhysicsBody::GetRotation(bool asRadians)
+    {
+        ZERO_VECTOR.Set(0, 0, 0);
+        if (!owner)
+            return ZERO_VECTOR;
+
+        return owner->GetTransform()->Rotation * (asRadians ? CE_DEG_TO_RAD : 1.0f);
+    }
 }