+++ GameCode/Test.cpp
@@ -11,14 +11,14 @@ void TestScript::OnUpdate()
         return;
 
     if (Input::GetKey(Keys::A))
-        rc->BodyInstance->AddForce({-1, 0, 0});
+        rc->BodyInstance->AddForce({-35, 0, 0});
 
     if (Input::GetKey(Keys::D))
-        rc->BodyInstance->AddForce({1, 0, 0});
+        rc->BodyInstance->AddForce({35, 0, 0});
 
     if (Input::GetKey(Keys::Q))
-        rc->BodyInstance->AddTorque({0, 0, 10});
+        rc->BodyInstance->AddTorque({0, 0, 360});
 
     if (Input::GetKey(Keys::R))
-        rc->BodyInstance->AddTorque({-10, 0, 0});
+        rc->BodyInstance->AddTorque({-360, 0, 0});
 }